Split Slack channel sync to its own commander
