+++ kifi-backend/modules/shoebox/test/com/keepit/test/ShoeboxInjectionHelpers.scala
@@ -95,6 +95,7 @@ trait ShoeboxInjectionHelpers { self: TestInjectorProvider =>
   def slackCommander(implicit injector: Injector) = inject[SlackIdentityCommander].asInstanceOf[SlackIdentityCommanderImpl]
   def slackIntegrationCommander(implicit injector: Injector) = inject[SlackIntegrationCommander].asInstanceOf[SlackIntegrationCommanderImpl]
   def slackTeamCommander(implicit injector: Injector) = inject[SlackTeamCommander].asInstanceOf[SlackTeamCommanderImpl]
+  def slackChannelCommander(implicit injector: Injector) = inject[SlackTeamCommander].asInstanceOf[SlackChannelCommander]
   def slackInfoCommander(implicit injector: Injector) = inject[SlackInfoCommander].asInstanceOf[SlackInfoCommanderImpl]
   def libToSlackPusher(implicit injector: Injector) = inject[LibraryToSlackChannelPusher].asInstanceOf[LibraryToSlackChannelPusherImpl]
   def slackClient(implicit injector: Injector) = inject[SlackClient].asInstanceOf[FakeSlackClientImpl]
